Alert

+++ components/FavoriteComponent.js
@@ -1,7 +1,7 @@
 import React, { Coponent, Component } from 'react';
 import { connect } from 'react-redux';
 import { Loading } from './LoadingComponent';
-import { View, Text, FlatList } from 'react-native';
+import { View, Text, FlatList, Alert } from 'react-native';
 import { ListItem } from 'react-native-elements';
 import { baseUrl } from '../shared/baseURL';
 import Swipeout from 'react-native-swipeout';
@@ -30,7 +30,21 @@ class Favorites extends Component {
                 {
                     text: 'Delete',
                     type: 'delete',
-                    onPress: () => this.props.deleteFavorite(item.id)
+                    onPress: () => {
+                        Alert.alert('Delete Favorite?', 
+                                    'Are you sure you want to delete the favorite dish ' + item.name + '?', 
+                                    [
+                                        {
+                                            text: 'Cancel',
+                                            onPress: () => console.log(item.name + 'Not Deleted'),
+                                            style: 'cancel'
+                                        },
+                                        {
+                                            text: 'ok',
+                                            onPress: () => this.props.deleteFavorite(item.id)
+                                        }
+                                    ], { cancelable: false });    
+                    }
                 }
             ];
